refactor(server): reorganize rest api services
