--- server/blueprint/index.js
@@ -1,5 +1,0 @@
-'use strict'
-
-const Blueprint = require('./blueprint')
-
-module.exports = new Blueprint()
