+++ server/xively/index.js
@@ -1,0 +1,11 @@
+'use strict'
+
+const blueprint = require('./blueprint')
+const idm = require('./idm')
+const integration = require('./integration')
+
+module.exports = {
+  blueprint,
+  idm,
+  integration
+}
